fix(sdk): harden websocket disconnect reason handling
Persist terminal disconnect reasons across cleanup and route upgrade parsing through the real Phoenix initial-connection boundary.
