+++ src/thenvoi/platform/link.py
@@ -99,6 +99,9 @@ class ThenvoiLink:
         # Event queue for async iteration
         self._event_queue: asyncio.Queue[PlatformEvent] = asyncio.Queue(maxsize=1000)
 
+        # Durable terminal disconnect reason for the current connection lifecycle.
+        self._last_disconnect_reason: WebSocketDisconnectReason | None = None
+
     @property
     def is_connected(self) -> bool:
         return self._is_connected
@@ -106,9 +109,7 @@ class ThenvoiLink:
     @property
     def last_disconnect_reason(self) -> WebSocketDisconnectReason | None:
         """Most recent terminal WebSocket disconnect reason, if reported."""
-        if self._ws is None:
-            return None
-        return self._ws.last_disconnect_reason
+        return self._last_disconnect_reason
 
     # --- Async iterator protocol ---
 
@@ -132,6 +133,7 @@ class ThenvoiLink:
             logger.warning("Already connected")
             return
 
+        self._last_disconnect_reason = None
         self._ws = WebSocketClient(
             self.ws_url,
             self.api_key,
@@ -313,6 +315,7 @@ class ThenvoiLink:
     async def _on_supersede(self, payload: "SupersedePayload") -> None:
         """Handle terminal supersede event before the platform closes the socket."""
         reason = payload.to_disconnect_reason()
+        self._last_disconnect_reason = reason
         self._is_connected = False
         if self._ws:
             self._ws.record_terminal_disconnect(reason)
